update
I finished the code for analyzing the logfile data it now stores the associated ip the error code or the request along side the http code and the suspicion value.
it currently uses test data.
it currently does not run in the main environment but has been tested in an exclusive environment.
i used a struct for testing purposes this can be removed if not needed later.
unsure if the hasmap works currently.